+++ examples/example_filters.cpp
@@ -18,6 +18,7 @@ public:
                               quill::detail::LogRecordMetadata const& metadata,
                               fmt::memory_buffer const& formatted_record) noexcept override
   {
+    // log only messages that are INFO or above in the file
     return metadata.level() < quill::LogLevel::Warning;
   }
 };
@@ -34,6 +35,7 @@ public:
                               quill::detail::LogRecordMetadata const& metadata,
                               fmt::memory_buffer const& formatted_record) noexcept override
   {
+    // log only WARNINGS or higher to stdout
     return metadata.level() >= quill::LogLevel::Warning;
   }
 };