+++ quill/include/quill/handlers/Handler.h
@@ -11,6 +11,7 @@
 #include "quill/detail/misc/Common.h"
 #include "quill/detail/misc/RecursiveSpinlock.h"
 #include "quill/filters/FilterBase.h"
+#include <atomic>
 #include <memory>
 #include <vector>
 
@@ -75,27 +76,35 @@ public:
   QUILL_ATTRIBUTE_HOT virtual void flush() noexcept = 0;
 
   /**
-   * Sets a log level filter on the handler. Log statements with higher or equal severity only will be loged
+   * Sets a log level filter on the handler. Log statements with higher or equal severity only will be logged
+   * @note thread safe
    * @param log_level the log level severity
    */
-  void set_log_level(LogLevel log_level);
+  void set_log_level(LogLevel log_level) { _log_level.store(log_level, std::memory_order_relaxed); }
 
   /**
    * Looks up the existing log level filter that was set by set_log_level and returns the current log level
-   * @return the current log level of the log level filter - if any
+   * @note thread-safe
+   * @return the current log level of the log level filter
    */
-  QUILL_NODISCARD LogLevel get_log_level() noexcept;
+  QUILL_NODISCARD LogLevel get_log_level() const noexcept
+  {
+    return _log_level.load(std::memory_order_relaxed);
+  }
 
   /** Filters **/
 
   /**
-   * Adds a new filter for this handler
+   * Adds a new filter for this handler.
+   * Filters can be added at any time to the handler.
+   * @note: thread-safe
    * @param filter instance of a filter class as unique ptr
    */
   void add_filter(std::unique_ptr<FilterBase> filter);
 
   /**
-   * Apply all registered filters
+   * Apply all registered filters.
+   * @note: called internally by the backend worker thread.
    * @return result of all filters
    */
   QUILL_NODISCARD bool apply_filters(char const* thread_id, std::chrono::nanoseconds log_record_timestamp,
@@ -117,7 +126,7 @@ private:
   /** Indicator that a new filter was added **/
   alignas(detail::CACHELINE_SIZE) std::atomic<bool> _new_filter{false};
 
-  quill::LogLevel _log_level;
+  std::atomic<quill::LogLevel> _log_level{LogLevel::TraceL3};
 };
 
 } // namespace quill